requirements.txt
altgraph==0.17.4
importlib_metadata==7.1.0
packaging==24.0
pefile==2023.2.7
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
python-vlc==3.0.20123
pywin32-ctypes==0.2.2
tk==0.1.0
zipp==3.18.1
